Conflicts:
	api/Application.class.php
	features/cerberusweb.core/api/dao/address.php
	features/cerberusweb.core/api/dao/contact_org.php
	features/cerberusweb.core/api/dao/task.php
	features/cerberusweb.core/api/dao/ticket.php
	features/cerberusweb.core/api/events/abstract/abstract_event_address.php
	features/cerberusweb.core/api/events/abstract/abstract_event_message.php
	features/cerberusweb.core/api/events/abstract/abstract_event_org.php
	features/cerberusweb.core/api/events/abstract/abstract_event_task.php
	features/cerberusweb.core/api/events/abstract/abstract_event_ticket.php
	features/cerberusweb.core/api/uri/internal.php
	features/cerberusweb.core/patches/5.x/5.8.0.php
	features/cerberusweb.core/plugin.xml
	features/cerberusweb.core/strings.xml
	features/cerberusweb.core/templates/configuration/index.tpl
	features/cerberusweb.core/templates/internal/decisions/actions/_create_comment.tpl
	features/cerberusweb.core/templates/internal/decisions/actions/_create_notification.tpl
	features/cerberusweb.core/templates/internal/decisions/actions/_set_var_list.tpl
	features/cerberusweb.core/templates/internal/decisions/actions/_shared_var_picker.tpl
	features/cerberusweb.core/templates/internal/decisions/actions/_shared_var_worker_picker.tpl
	features/cerberusweb.core/templates/internal/views/criteria/__list.tpl
	features/cerberusweb.core/templates/tickets/search/criteria/ticket_status.tpl
	features/cerberusweb.crm/api/dao/crm_opportunity.php
	features/cerberusweb.crm/api/events/crm_opportunity_macro.php
	features/cerberusweb.kb/api/dao/kb_article.php
	features/cerberusweb.kb/api/events/abstract/abstract_event_kbarticle.php
	features/cerberusweb.timetracking/api/dao/timetracking_entry.php
	features/cerberusweb.timetracking/api/events/abstract_event_timetracking.php
	libs/devblocks/api/Extension.php

+++ features/cerberusweb.core/api/uri/internal.php
@@ -190,8 +190,35 @@ class ChInternalController extends DevblocksControllerExtension {
 		@$context = DevblocksPlatform::importGPC($_REQUEST['context'],'string');
 		@$view_id = DevblocksPlatform::importGPC($_REQUEST['view_id'],'string');
 		@$layer = DevblocksPlatform::importGPC($_REQUEST['layer'],'string');
-
+		@$trigger_id = DevblocksPlatform::importGPC($_REQUEST['trigger_id'],'integer',0);
+		
 		if(null != ($view = C4_AbstractViewLoader::getView($view_id))) {
+			if(!empty($trigger_id) && null != ($trigger = DAO_TriggerEvent::get($trigger_id))) {
+				$event = $trigger->getEvent();
+				
+				if(method_exists($event,'generateSampleEventModel')) {
+					$event_model = $event->generateSampleEventModel();
+					$event->setEvent($event_model);
+					$values = $event->getValues();
+					$view->setPlaceholderValues($values);
+				}
+				
+				$conditions = $event->getConditions($trigger);
+				$valctx = $event->getValuesContexts($trigger);
+				foreach($valctx as $token => $vtx) {
+					$conditions[$token] = $vtx;
+				}
+
+				foreach($conditions as $cond_id => $cond) {
+					if(substr($cond_id,0,1) == '_')
+						unset($conditions[$cond_id]);
+				}
+				
+				$view->setPlaceholderLabels($conditions);
+				
+				C4_AbstractViewLoader::setView($view->id, $view);
+			}
+			
 			$tpl = DevblocksPlatform::getTemplateService();
 			$tpl->assign('context', $context);
 			$tpl->assign('layer', $layer);
@@ -984,38 +1011,51 @@ class ChInternalController extends DevblocksControllerExtension {
 
 	function viewRefreshAction() {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
-		$view = C4_AbstractViewLoader::getView($id);
-		$view->render();
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			$view->render();
+		}
 	}
 
 	function viewSortByAction() {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 		@$sortBy = DevblocksPlatform::importGPC($_REQUEST['sortBy']);
 
-		$view = C4_AbstractViewLoader::getView($id);
-		$view->doSortBy($sortBy);
-		C4_AbstractViewLoader::setView($id, $view);
-
-		$view->render();
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			$view->doSortBy($sortBy);
+			C4_AbstractViewLoader::setView($id, $view);
+			$view->render();
+		}
 	}
 
 	function viewPageAction() {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 		@$page = DevblocksPlatform::importGPC(DevblocksPlatform::importGPC($_REQUEST['page']));
 
-		$view = C4_AbstractViewLoader::getView($id);
-		$view->doPage($page);
-		C4_AbstractViewLoader::setView($id, $view);
-
-		$view->render();
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			$view->doPage($page);
+			C4_AbstractViewLoader::setView($id, $view);
+			$view->render();
+		}
 	}
 
 	function viewGetCriteriaAction() {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 		@$field = DevblocksPlatform::importGPC($_REQUEST['field']);
 
-		$view = C4_AbstractViewLoader::getView($id);
-		$view->renderCriteria($field);
+		$tpl = DevblocksPlatform::getTemplateService();
+		
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			$tpl->assign('view', $view);
+			
+			// Do we already have this filter to re-edit?
+			$params = $view->getEditableParams();
+			if(isset($params[$field])) {
+				$tpl->assign('param', $params[$field]);
+			}
+			
+			// Render from the View_* implementation.
+			$view->renderCriteria($field);
+		}
 	}
 
 	private function _viewRenderInlineFilters($view, $is_custom=false) {
@@ -1034,9 +1074,10 @@ class ChInternalController extends DevblocksControllerExtension {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id'],'string','');
 		@$show = DevblocksPlatform::importGPC($_REQUEST['show'],'integer',0);
 		
-		$view = C4_AbstractViewLoader::getView($id);
-		$view->renderFilters = !empty($show) ? 1 : 0;
-		C4_AbstractViewLoader::setView($view->id, $view);
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			$view->renderFilters = !empty($show) ? 1 : 0;
+			C4_AbstractViewLoader::setView($view->id, $view);
+		}
 	}
 	
 	function viewAddFilterAction() {
@@ -1048,7 +1089,8 @@ class ChInternalController extends DevblocksControllerExtension {
 		@$value = DevblocksPlatform::importGPC($_REQUEST['value']);
 		@$field_deletes = DevblocksPlatform::importGPC($_REQUEST['field_deletes'],'array',array());
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
 
 		if($is_custom && 0 != strcasecmp('cust_',substr($id,0,5)))
 			$is_custom = 0;
@@ -1085,7 +1127,8 @@ class ChInternalController extends DevblocksControllerExtension {
 	function viewResetFiltersAction() {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
 
 		$view->doResetCriteria();
 
@@ -1098,7 +1141,8 @@ class ChInternalController extends DevblocksControllerExtension {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 		@$preset_id = DevblocksPlatform::importGPC($_REQUEST['_preset'],'integer',0);
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
 
 		$view->removeAllParams();
 
@@ -1122,7 +1166,9 @@ class ChInternalController extends DevblocksControllerExtension {
 
 		$active_worker = CerberusApplication::getActiveWorker();
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
+		
 		$params_json = json_encode($view->getEditableParams());
 
 		if(!empty($preset_replace_id)) {
@@ -1159,11 +1205,10 @@ class ChInternalController extends DevblocksControllerExtension {
 		@$id = DevblocksPlatform::importGPC($_REQUEST['id']);
 		@$preset_dels = DevblocksPlatform::importGPC($_REQUEST['_preset_del'],'array',array());
 
-		$view = C4_AbstractViewLoader::getView($id);
-
-		DAO_ViewFiltersPreset::delete($preset_dels);
-
-		$this->_viewRenderInlineFilters($view);
+		if(null != ($view = C4_AbstractViewLoader::getView($id))) {
+			DAO_ViewFiltersPreset::delete($preset_dels);
+			$this->_viewRenderInlineFilters($view);
+		}
 	}
 
 	function viewCustomizeAction() {
@@ -1174,7 +1219,8 @@ class ChInternalController extends DevblocksControllerExtension {
 		$tpl = DevblocksPlatform::getTemplateService();
 		$tpl->assign('id', $id);
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
 		
 		// Custom worklists
 		if('cust_' == substr($view->id,0,5)) {
@@ -1215,7 +1261,8 @@ class ChInternalController extends DevblocksControllerExtension {
 		$active_worker = CerberusApplication::getActiveWorker();
 		$tpl = DevblocksPlatform::getTemplateService();
 
-        $view = C4_AbstractViewLoader::getView($view_id);
+        if(null == ($view = C4_AbstractViewLoader::getView($view_id)))
+        	return;
 
 		$workspaces = DAO_Workspace::getByOwner(CerberusContexts::CONTEXT_WORKER, $active_worker->id);
 		$tpl->assign('workspaces', $workspaces);
@@ -1232,7 +1279,9 @@ class ChInternalController extends DevblocksControllerExtension {
 		$active_worker = CerberusApplication::getActiveWorker();
 
 	    @$view_id = DevblocksPlatform::importGPC($_POST['view_id'],'string');
-		$view = C4_AbstractViewLoader::getView($view_id);
+	    
+		if(null == ($view = C4_AbstractViewLoader::getView($view_id)))
+			return;
 
 		@$list_title = DevblocksPlatform::importGPC($_POST['list_title'],'string', '');
 		@$workspace_id = DevblocksPlatform::importGPC($_POST['workspace_id'],'integer', 0);
@@ -1293,7 +1342,8 @@ class ChInternalController extends DevblocksControllerExtension {
 		$tpl = DevblocksPlatform::getTemplateService();
 		$tpl->assign('view_id', $view_id);
 
-		$view = C4_AbstractViewLoader::getView($view_id);
+		if(null == ($view = C4_AbstractViewLoader::getView($view_id)))
+			return;
 		$tpl->assign('view', $view);
 
 		$model_columns = $view->getColumnsAvailable();
@@ -1317,7 +1367,9 @@ class ChInternalController extends DevblocksControllerExtension {
 				unset($columns[$idx]);
 		}
 
-		$view = C4_AbstractViewLoader::getView($view_id);
+		if(null == ($view = C4_AbstractViewLoader::getView($view_id)))
+			return;
+		
 		$column_manifests = $view->getColumnsAvailable();
 
 		// Override display
@@ -1421,7 +1473,9 @@ class ChInternalController extends DevblocksControllerExtension {
 			}
 		}
 
-		$view = C4_AbstractViewLoader::getView($id);
+		if(null == ($view = C4_AbstractViewLoader::getView($id)))
+			return;
+		
 		$view->doCustomize($columns, $num_rows);
 
 		$is_custom = substr($id,0,5)=='cust_';
@@ -2833,6 +2887,12 @@ class ChInternalController extends DevblocksControllerExtension {
 		$tpl->assign('output', $output);
 		$tpl->display('devblocks:cerberusweb.core::internal/renderers/test_results.tpl');
 	}
+	
+	// Scheduled Behavior
+	
+	function showScheduledBehaviorTabAction() {
+		Subcontroller_Internal_VirtualAttendants::showScheduledBehaviorAction();
+	}	
 	
 	// Calendars
